+++ packages/terra-core-docs/src/terra-dev-site/test/button/ActiveVariantButtons.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import IconSquare from 'terra-icon/lib/icon/IconSquareSymbol';
 import classNames from 'classnames/bind';
 import buttonStyles from 'terra-button/lib/Button.module.scss';
-import Button from '../../../Button';
+import Button from 'terra-button';
 import styles from './ButtonTestCommon.module.scss';
 
 const cx = classNames.bind(styles);