+++ packages/terra-core-docs/src/terra-dev-site/test/button/CompactButton.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../../../Button';
+import Button from 'terra-button';
 
 const CompactButton = () => <Button id="compactButton" text="Compact" isCompact />;