+++ packages/terra-core-docs/src/terra-dev-site/test/button/IconButton.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import IconSquare from 'terra-icon/lib/icon/IconSquareSymbol';
 import classNames from 'classnames/bind';
-import Button from '../../../Button';
+import Button from 'terra-button';
 import styles from './ButtonTestCommon.module.scss';
 
 const cx = classNames.bind(styles);