+++ packages/terra-core-docs/src/terra-dev-site/test/button/ProgrammaticDisableEnable.test.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '../../../Button';
+import Button from 'terra-button';
 
 const ProgrammaticDisableEnable = () => {
   const [isDisabled, setIsDisabled] = useState(false);